refactor: removes clap-derive-internals crate as it was redundant and moved derive traits into mainline clap
