--- clap-derive-internals/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "clap-derive-internals"
-description = "clap-custom-derive internals and procedural macros"
-version = "0.0.6"
-authors = [
-	"hoverbear <andrew@hoverbear.org>",
-	"Guillaume Pinot <texitoi@texitoi.eu>",
-	"Kevin K. <kbknapp@gmail.com>"
-]
-license = "MIT/Apache-2.0"
-categories = ["development-tools::procedural-macro-helpers"]
-keywords = ["macro", "clap", "command-line", "structopt", "derive"]
-readme = "README.md"
-repository = "https://github.com/kbknapp/clap-derive"
-
-[badges]
-travis-ci = { repository = "https://github.com/kbknapp/clap-derive" }
-appveyor = { repository = "https://github.com/kbknapp/clap-derive", service = "github" }
-
-[lib]
-proc-macro = true
-
-[dependencies]
-quote = "*"
-syn = "*"
-error-chain = "*"
-clap = "2"
-clippy    = { version = "~0.0.166", optional = true }
-
-[features]
-default     = []
-unstable    = [] # for building with unstable clap features (doesn't require nightly Rust) 
-nightly     = [] # for building with unstable Rust features 
-lints       = ["clippy"] # Requires nightly Rust
-
-[package.metadata.docs.rs]
-features = ["doc"]
